% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_epconstraint_obj2_approach.R
\name{add_epconstraint_obj2_approach}
\alias{add_epconstraint_obj2_approach}
\title{Add epsilon constraint approach for 2 objectives}
\usage{
add_epconstraint_obj2_approach(x, n, verbose = TRUE)
}
\arguments{
\item{x}{\code{\link[prioritizr:multi_problem]{prioritizr::multi_problem()}} object.}

\item{n}{\code{integer} number of solutions. Note that \code{n} must be greater
than or equal to the number of objectives in \code{x}.}

\item{verbose}{\code{logical} should progress on generating solutions
displayed? Defaults to \code{TRUE}.}
}
\value{
TODO.
}
\description{
Add an epsilon-constraint approach for multi-objective optimization to a
conservation planning problem.
}
\details{
TODO.
}
\section{Mathematical formulation}{

TODO.
}

\examples{
\dontrun{
# TODO
}

}
\seealso{
TODO.
}
