% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_epconstraint_approach.R
\name{add_epconstraint_approach}
\alias{add_epconstraint_approach}
\title{Add epsilon constraint approach}
\usage{
add_epconstraint_approach(x, n_per_problem, verbose = TRUE)
}
\arguments{
\item{x}{\code{\link[prioritizr:multi_problem]{prioritizr::multi_problem()}} object.}

\item{n_per_problem}{\code{integer} number of solutions to generate per
problem in \code{x}. Note that this does not include solutions that represent
optimizing exclusively on each objective individually (i.e., extreme points
of the Pareto frontier). In particular, the total number of solutions
is equal to \eqn{o + n^(o - 1)}{o + n^(o - 1)}, where
\eqn{o} is the number of problems in \code{x} and \eqn{n} is equal to
\code{n_per_problem}. For example, if \code{x} had three problems and
\code{n_per_problem = 4}, then 19 solutions would be generated.}

\item{verbose}{\code{logical} (\code{TRUE}/\code{FALSE}) should progress on generating
solutions displayed? Defaults to \code{TRUE}.}
}
\value{
An updated \code{\link[prioritizr:multi_problem]{prioritizr::multi_problem()}} object with the approach
added to it.
}
\description{
Add an epsilon-constraint approach for multi-objective optimization to a
conservation planning problem.
}
\details{
TODO.
}
\section{Mathematical formulation}{

TODO.
}

\examples{
\dontshow{if (prioritizr::do_run_example()) withAutoprint(\{ # examplesIf}
# in this example, we aim to identify a set of planning units that will
# not exceed a particular budget and meet objectives for
# (i) representing species that are important for ecosystem
# functioning (hereafter, keystone species) and (ii) representing species
# that have high social or cultural value (hereafter, iconic species)

# load packages
library(prioritizr)
library(terra)

# import data
con_cost <- get_sim_pu_raster()
keystone_spp <- get_sim_features()[[1:3]]
iconic_spp <- get_sim_features()[[4:5]]

# define a total conservation budget (20\% of total cost)
budget <- terra::global(con_cost, "sum", na.rm = TRUE)[[1]] * 0.2

# define a single-objective problem for the keystone species objective
p1 <-
  problem(con_cost, keystone_spp) \%>\%
  add_min_shortfall_objective(budget) \%>\%
  add_relative_targets(0.4) \%>\%
  add_binary_decisions()

# define a single-objective problem for the iconic species objective
p2 <-
  problem(con_cost, iconic_spp) \%>\%
  add_min_shortfall_objective(budget) \%>\%
  add_relative_targets(0.8) \%>\%
  add_binary_decisions()

# now create multi-objective problem with epsilon-constraint approach,
# with settings to generate (i) a solution for each objective
# and (ii) three additional solutions that represent a varying
# degree of trade-offs between these objectives
mp <-
  multi_problem(keystone_obj = p1, iconic_obj = p2) \%>\%
  add_epconstraint_approach(n_per_problem = 4, verbose = TRUE) \%>\%
  add_default_solver(gap = 0, verbose = FALSE)

# solve problem
ms <- solve(mp)

# convert solutions to a multi-layer raster object
sol <- rast(ms)

# assign names to each solution
names(sol) <- c(
  "Keystone species", "Iconic species",
  paste("trade-off", seq_len(nlyr(sol) - 2))
)

# plot solutions to view selection of priority areas,
# here, the first two maps show solutions for best achieving
# each of the objectives, and the subsequent maps show solutions
# that aim to reach varying degrees of compromise between the objectives
plot(sol, axes = FALSE)

# extract objective values for the solutions
obj_matrix <- attributes(ms)$objective

# print the objective values
print(obj_matrix)

# plot the objectives values to visualize trade-offs
# (note that smaller values are better because these objectives seek to
# minimize representation shortfalls)
plot(
  obj_matrix,
  main = "Trade-offs between objectives",
  xlab = "Keystone objective (shortfall)",
  ylab = "Iconic objective (shortfall)"
)
\dontshow{\}) # examplesIf}
}
\seealso{
See \link[prioritizr:approaches]{prioritizr::approaches} for other functions for adding an approach.
}
